updated access control service and moved to the common Token Payload interface
