+++ src/access-control/constants/access-control-error.enum.ts
@@ -2,5 +2,5 @@ export enum ACCESS_CONTROL_ERROR {
   CREATE_ACCESS_ERROR = 'Error create access control',
   UPDATE_ACCESS_ERROR = 'Error update access control',
   DELETE_ACCESS_ERROR = 'Error delete user',
-  DUPLICATE_ACCESS_ERROR = 'Access control already exists',
+  OWN_ACCESS_ERROR = 'Are you trying to add access to your expenses',
 }
