+++ src/common/interfaces/token.interface.ts
